refactor dry concept
