+++ src/commands/kang.ts
@@ -17,6 +17,7 @@ import { IndexPackInvalid } from '../error/IndexPackInvalid';
 import { KangInputInvalid } from '../error/KangInputInvalid';
 import { StickerTooMuchSynteticTrue } from '../error/stickerTooMuch';
 import { logger } from '../utils/logger';
+import { dbHelper } from '../helper/dbHelper';
 
 interface tgCallErrcode {
     code: number;
@@ -67,45 +68,6 @@ class stickerPackManagement {
         let highestIdx = await this.findCurrentHighestIdx(stickerNameArr);
         let generatedName: IStickerpackData = await BotHelpers.genStickerpackName(this.ctx, highestIdx);
         return generatedName.stickerName
-    }
-
-    private async appendNewStickerSetToUID(user_id: number, newStickersetName: string) {
-        let old = await stickerpackStateModel.findOne({
-            user_id: user_id
-        });
-
-        let newAppendSticker: string[] = old?.stickersetname!;
-        newAppendSticker.push(newStickersetName!);
-
-        await stickerpackStateModel.updateOne(
-            {
-                user_id: user_id
-            },
-            {
-                stickersetname: newAppendSticker
-            }
-        )
-    }
-
-    private async removeStickerSetFromUID(user_id: number, stickersetName: string) {
-        let old = await stickerpackStateModel.findOne({
-            user_id: user_id
-        });
-
-        let repackArr: string[] = old?.stickersetname!;
-        const index = repackArr.indexOf(stickersetName);
-        if (index > -1) { // only splice array when item is found
-            repackArr.splice(index, 1); // 2nd parameter means remove one item only
-        }
-
-        await stickerpackStateModel.updateOne(
-            {
-                user_id: user_id
-            },
-            {
-                stickersetname: repackArr
-            }
-        )
     }
 
     private async addStickerToPack(stickerFileId: string):
@@ -395,7 +357,7 @@ class stickerPackManagement {
                 /**
                  * create completion
                  */
-                await this.appendNewStickerSetToUID(
+                await dbHelper.appendNewStickerSetToUID(
                     this.ctx.message?.from.id!,
                     this.data.stickerName!
                 );
@@ -409,14 +371,14 @@ class stickerPackManagement {
                 /**
                  * malformed data, need to delete
                  */
-                await this.removeStickerSetFromUID(
+                await dbHelper.removeStickerSetFromUID(
                     this.ctx.message?.from.id!,
                     this.data.stickerName!
                 )
             } else if (tgServerSideFoundSticker == false && localFoundSticker == false) {
                 ret = await this.createNewStickerpack(stickerFileId);
                 if (ret.code == 0) {
-                    await this.appendNewStickerSetToUID(
+                    await dbHelper.appendNewStickerSetToUID(
                         this.ctx.message?.from.id!,
                         this.data.stickerName!
                     );
